rename _AIX_platform to _aix_support - same as in CPython
